+++ src/Mqtt5ClientFilterConfigWidget.cpp
@@ -145,7 +145,7 @@ void Mqtt5ClientFilterConfigWidget::refresh()
         addTopicAliasWidget(aliasConfig);
     }
 
-    m_ui.m_respTimeoutSpinBox->setValue(m_filter.config().m_respTimeout);
+    m_ui.m_respTimeoutSpinBox->setValue(static_cast<int>(m_filter.config().m_respTimeout));
     m_ui.m_clientIdLineEdit->setText(m_filter.config().m_clientId);
     m_ui.m_usernameLineEdit->setText(m_filter.config().m_username);
     m_ui.m_passwordLineEdit->setText(m_filter.config().m_password);
